Start building components and getting them on screen
