+++ gtk/src/components/mod.rs
@@ -1,0 +1,5 @@
+mod day;
+mod time_distance;
+
+pub use day::day_c;
+pub use time_distance::time_distance_c;
